app/requirements.txt
fastapi[all]
jinja2
jinja2-fragments
python-multipart
pytest
pytailwindcss
tinydb
uvicorn
pandas
openai
sqlalchemy
simplejson
psycopg2
python-dotenv
python-dateutil
beautifulsoup4
paho-mqtt
sse_starlette
asyncio